+++ src/main/java/com/programmers/devcourse/baseball/custom/exception/NotProperNumber.java
@@ -1,4 +1,4 @@
-package com.programmers.devcourse.baseball.exception;
+package com.programmers.devcourse.baseball.custom.exception;
 
 public class NotProperNumber extends RuntimeException{
 	public NotProperNumber(String message) {
