+++ src/main/java/com/programmers/devcourse/baseball/custom/exception/NotUniqueNumber.java
@@ -1,4 +1,4 @@
-package com.programmers.devcourse.baseball.exception;
+package com.programmers.devcourse.baseball.custom.exception;
 
 public class NotUniqueNumber extends RuntimeException {
 	public NotUniqueNumber(String message) {
